requirements.txt
colorzero==2.0
gpiozero==1.6.2
numpy==1.21.1
opencv-python==3.4.6.27
picamera==1.13
Pillow==8.3.1
pkg-resources==0.0.0
RPi.GPIO==0.7.0
tflite-runtime==2.5.0
